requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pyodbc==4.0.39
pytest==7.3.1
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.14
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
